+++ src/pii_transform/api/e2e/defs.py
@@ -1,0 +1,3 @@
+
+# Minimum vesion  required of pii-extract-base package
+MIN_PII_EXTRACT_VERSION = "0.4.0"