+++ test/unit/api/test_textchunk_multi.py
@@ -6,7 +6,10 @@ from pathlib import Path
 
 from pii_data.types.doc import DocumentChunk
 from pii_data.types.piicollection.loader import PiiCollectionLoader
+from pii_data.types.piicollection import PiiDetector
 from pii_data.helper.exception import ProcException
+
+import pii_extract.api.processor as processor
 
 import pii_transform.api.e2e.multilang as mod
 
@@ -16,21 +19,27 @@ import pytest
 DATADIR = Path(__file__).parents[2] / "data" / "chunk"
 
 
-def patch_pii_extract(monkeypatch, piic=None):
+def patch_pii_extract(monkeypatch,
+                      piic_result: processor.PiiCollectionBuilder = None):
     """
     Patch the pii-extract-base API used by textchunk
     """
+
+    def detect(chunk, piic):
+        piic.add_collection(piic_result)
+
     processor_mock = Mock()
-    processor_mock.get_stats = Mock()
     processor_mock.build_tasks = Mock(return_value=2)
+    processor_mock.detect_chunk = detect
+    processor_mock.get_stats = Mock(return_value={})
     processor_cls = Mock(return_value=processor_mock)
-    collection_mock = Mock(return_value=piic)
+    #collection_mock = Mock(return_value=piic)
     # Remove the mark of import fail
-    monkeypatch.setattr(mod, "MISSING", None)
+    monkeypatch.setattr(mod, "MISSING_MOD", None)
     # Set the processor to be a mock
-    monkeypatch.setattr(mod, "PII_EXTRACT_VERSION", "0.3.1")
+    monkeypatch.setattr(mod, "PII_EXTRACT_VERSION", "0.4.0")
     monkeypatch.setattr(mod, "PiiProcessor", processor_cls)
-    monkeypatch.setattr(mod, "PiiCollectionBuilder", collection_mock)
+    #monkeypatch.setattr(mod, "PiiCollectionBuilder", collection_mock)
     return processor_mock
 
 
@@ -49,7 +58,6 @@ def test10_constructor(monkeypatch):
     assert mock.build_tasks.call_count == 2
 
 
-
 def test11_constructor_error(monkeypatch):
     """
     Test constructing the object, error
@@ -64,9 +72,10 @@ def test20_process_chunk(monkeypatch):
     """
     Test processing a text buffer
     """
-    piic = PiiCollectionLoader()
-    piic.load_json(DATADIR / "example-pii.json")
-    patch_pii_extract(monkeypatch, piic)
+    # Prepare the data that will be reported by the extractor
+    piic1 = PiiCollectionLoader()
+    piic1.load_json(DATADIR / "example-pii.json")
+    patch_pii_extract(monkeypatch, piic1)
 
     with open(DATADIR / "example-src.txt", encoding="utf-8") as f:
         src = f.read()
@@ -80,7 +89,7 @@ def test20_process_chunk(monkeypatch):
     got = m(src, "en")
     assert exp == got
 
-    # We get the same result, since we are patching detection
+    # We get the same result for "es", since we have patched detection
     got = m(src, "es")
     assert exp == got
 
@@ -127,25 +136,81 @@ def test50_process_chunk_err(monkeypatch):
 
 def test60_process_piic(monkeypatch):
     """
-    Test stats
+    Test keep_piic
     """
-    piic = PiiCollectionLoader()
-    piic.load_json(DATADIR / "example-pii.json")
-    patch_pii_extract(monkeypatch, piic)
+    # Load the data that will be "extracted"
+    piic1 = PiiCollectionLoader()
+    piic1.load_json(DATADIR / "example-pii.json")
 
-    with open(DATADIR / "example-src.txt", encoding="utf-8") as f:
-        src = f.read()
+    # Create a collection builder, fill it with the data, and set it as
+    # the result of the extraction process
+    piic2 = processor.PiiCollectionBuilder()
+    piic2.add_collection(piic1)
+    patch_pii_extract(monkeypatch, piic2)
+
 
     m = mod.MultiPiiTextProcessor(["en", "es"], default_policy="annotate",
                                   keep_piic=True)
 
+    with open(DATADIR / "example-src.txt", encoding="utf-8") as f:
+        src = f.read()
     m(src, "en")
 
     piic = m.piic()
     assert len(piic) == 2
 
 
-def test60_process_stats(monkeypatch):
+def test61_process_piic_nolang(monkeypatch):
+    """
+    Test keep_piic, non-existing lang
+    """
+    piic1 = PiiCollectionLoader()
+    piic1.load_json(DATADIR / "example-pii.json")
+    piic2 = processor.PiiCollectionBuilder()
+    piic2.add_collection(piic1)
+    patch_pii_extract(monkeypatch, piic2)
+
+    with open(DATADIR / "example-src.txt", encoding="utf-8") as f:
+        src = f.read()
+
+    m = mod.MultiPiiTextProcessor(["en", "es", "zz"], default_policy="annotate",
+                                  keep_piic=True)
+    m(src, "en")
+
+    m(src, "zz")
+
+    piic = m.piic()
+    assert len(piic) == 4
+
+
+def test70_process_detectors(monkeypatch):
+    """
+    Test fetching the detectors
+    """
+    # Load the data that will be "extracted"
+    piic1 = PiiCollectionLoader()
+    piic1.load_json(DATADIR / "example-pii.json")
+
+    # Create a collection builder, fill it with the data, and set it as
+    # the result of the extraction process
+    piic2 = processor.PiiCollectionBuilder()
+    piic2.add_collection(piic1)
+    patch_pii_extract(monkeypatch, piic2)
+
+
+    m = mod.MultiPiiTextProcessor(["en", "es"], default_policy="annotate",
+                                  keep_piic=True)
+
+    with open(DATADIR / "example-src.txt", encoding="utf-8") as f:
+        src = f.read()
+    m(src, "en")
+
+    detectors = m.detectors()
+    assert len(detectors) == 1
+    assert detectors == {1: PiiDetector("PIISA", "PII Mock detector", "0.1.0")}
+
+
+def test80_process_stats(monkeypatch):
     """
     Test stats
     """
@@ -157,7 +222,9 @@ def test60_process_stats(monkeypatch):
         src = f.read()
 
     m = mod.MultiPiiTextProcessor(["en", "es"], default_policy="annotate")
-
     m(src, "en")
-    m.stats()
+
+    r = m.stats()
+    assert sorted(r.keys()) == ["chunks", "entities", "num", "time"]
+    assert r["chunks"] == {"total": {"en": 1}, "pii": {"en": 1}}
     assert proc.get_stats.called